AgePuzzle

Description

A module that solves the Age Problem Puzzle.

Mr. A tells a Mr. B that he has three sons. Mr. A then tells Mr. B that he will give him facts about his chldren until he has enough information to determine their ages. Mr. A tells Mr. B to stop him when Mr. B has enough information.

Mr. A: The sum of my boys' ages is 13.

Mr. A: The product of their ages is your age.

Mr. A: My eldest son lives in the attic.

Mr. B: Stop, I know the ages of your sons.

Question: What are the ages of the three sons?

Problem Solution: The code solves this problem more generically via the command line: ageProb sum-of-ages The solution to the original problem is (9,2,2).

Documentation

type Children = [Int]

Type Children is a list of ages.

main :: IO ()

Have user input the sum of the three sons ages, this program will provide all solutions.

kids :: (Num a, Enum a, Eq a) => a -> [[a]]

List all possible children in canonical order (oldest to youngest) which sum to n.

prod :: Children -> Int

Product of kids ages.

kpsort :: Int -> [Children]

Produce sorted children by the product of their ages.

kpgroup :: [Children] -> [[Children]]

Group children by equal age products.

ambprods :: [[Children]] -> [[Children]]

Filter for lists of length > 1 These are the ambiguous lists; otherwise Mr. B would know the answer and he needed more info to determine the answer.

oldestSon :: Children -> Bool

Check a child list for an eldest son.

psolns :: Int -> [[Children]]

Get all potential solutions.

Group the potential children by common age product.

Filter for groups of length > 1 (Mr. B needed more info)

Within the groups, filter out potential solutions that don't have an eldest son.

solns :: Int -> [Children]

Filter potential solutions to get final solutions. From the list of potential solutions, filter for singleton lists -- as these are the ones which allow Mr. B to know the answer -- and then concatenate them together to get all possible solutions.