Database

Description

Contém todos os tipos de dados que compõem o jogo.

Documentation

tam :: Float

Tamanho do lado de um bloco de peça.

limiteX :: Float

Limites do jogo.

limiteY :: Float

Limites do jogo.

type Coord = (Float, Float)

Uma coordenada (x, y).

data Tetramino

Tipo da peça baseado no Tetraminó.

Constructors

I

L

J

T

O

Z

S

data Peca

Cada peça é composta por um tipo (Tetraminó), uma cor, um foco e os blocos que a compõem.

Constructors

Peca

Fields

tipo :: Tetramino

cor :: Color

foco :: Coord

blocosPeca :: [Coord]

data Mundo

Um mundo é composto por seus blocos, uma peça ativa, uma lista de próximas peças, pontuação do jogo, um timer para aplicar a gravidade na peça, a quantidade de fileiras que foram completas na jogada, e uma flag de fim de jogo.

Constructors

Mundo

Fields

blocosPintados :: [(Coord, Color)]

pecaAtual :: Peca

proxPecas :: [Tetramino]

pontuacao :: Int

timer :: Int

fileirasCompletas :: Int

gameOver :: Bool

data Direcao

A direção do movimento pode ser para esquerda, para direita, para baixo ou rotação da peça.

Constructors

Dir

Esq

Baixo

Rotaciona

criaPeca :: Tetramino -> Peca

Cria uma peça nova de acordo com o tipo.