GA

Description

General implementation of GA for the knapsack problem.

Documentation

populationSize :: Int

Default population size.

mutationChance :: Double

Default mutation chance.

data GA_Args

Data type to store values that are not chaning thoughout the algorithm.

Constructors

Args [(Int, Int)] Int Double ([(Int, Int)] -> (Chromosome, Chromosome) -> IO Chromosome)

train :: String -> String -> Int -> Int -> IO Chromosome

Main training function.

Sorts each generation and starts generating new one if amount of generations has not been reached.

firstGen :: [(Int, Int)] -> Int -> Int -> IO [Chromosome]

Generating first random generation.

generateNextGen :: GA_Args -> Int -> Int -> IO [Chromosome] -> IO [Chromosome]

Generating next generation using crossover and mutation