Game

Documentation

data Player

Constructors

PlayerO1

PlayerO2

type Cell = Maybe Player

data State

Constructors

Running

GameOver (Maybe Player)

type Board = Array (Int, Int) Cell

data Game

Constructors

Game

Fields

gameBoard :: Board

gamePlayer :: Player

gameState :: State

n :: Int

screenWidth :: Int

Sets the ScreenWidth to 720

screenHeight :: Int

Sets the ScreenHeight to 720

cellWidth :: Float

Sets the Width of each cell corresponding to Screen width

cellHeight :: Float

Sets the height of each cell correspnding to Screen Height

initialGame :: Game

Sets the Initial Game State