Knapsack

Description

Implementation of needed functions for GA.

Documentation

newChromosome

Arguments

:: [(Int, Int)]
Database of (weight, value).

-> Int
Maximum amount of weight allowed.

-> IO Chromosome
New random chromosome.

Generate ONE new random solution.

This adds random items until weight restriction is met.

loadDatabaseFrom

Arguments

:: String
Relative path to file containing weights.

-> String
Relative path to file containing values.

-> IO [(Int, Int)]
Database of (weight, value) for each item.

Load database from weight and value text files.

Expected format: each file contains same amount of lines in both files and each line contains one positive integer. Function DOES NOT check the value of integers!

defaultCrossover

Arguments

:: [(Int, Int)]
Database of (weight, value) for each item.

-> (Chromosome, Chromosome)
Parents.

-> IO Chromosome
New chromosome created from parents.

Default implementation of crossover between 2 chromosomes.

Takes half (random set of indices) of items as chosen in first parent, rest from second parent. Does NOT check weight restriction.

defaultChooseParent

Arguments

:: [Chromosome]
Generation of Chromosome to choose from.

-> IO Chromosome
Chosen parent.

Default implementation of choosing parent from last generation.

Choosing random chromosome with higher value than random fraction of best solution. This algorithm prefers better solutions but can also choose worse ones.

getRandInts

Arguments

:: Int
Amount of random Ints to be generated.

-> Int
Max index.

-> IO [Int]
Unsorted list of non-repeating Ints

Get N random non-repeating unsorted Ints in range [0,maxIndex - 1]

If amount > maxIndex, returns pure []

>>> getRandInts 5 10
[3,7,1,5,4]

mutate

Arguments

:: [(Int, Int)]
Database of (weight, value) for each item.

-> Chromosome
Chromosome to mutate.

-> IO Chromosome
Mutated chromosome.

Implementation of mutation of a given chromosome.

Changes decision for ONE random item in the given chromosome.

getRandNum

Arguments

:: (Int, Int)
Interval [low, hi]

-> IO Int
Random int in interval

Get random int in range [lo, hi].

Does NOT check if lo <= hi.

getRandDouble

Arguments

:: (Double, Double)
Interval [low, hi]

-> IO Double
Random Double in interval

Get random double in range [lo, hi].

sortWith

Arguments

:: (a -> a -> Bool)
Comparing function.

-> [a]
List to sort.

-> [a]
Sorted list.

Sorts a list using merge sort algorithm.

data Chromosome

Type used to store a solution.

Stores weight, value and list of items (True -> chosen, False -> NOT chosen)

Constructors

Chromosome Int Int [Bool]