Logic

Documentation

isCoordinateCorrect :: (Int, Int) -> Bool

isCoordinateCorrect : This checks if the integer pair returned from mousePositionAsCellCoordinate are indeed on the board.

switchPlayer :: Game -> Game

switchPlayer : This switches the player after completion of a turn.

full :: [Cell] -> Maybe Player

full : This groups all the buttons of a player after every turn to check whether the player can satisfy the winning criteria.

winner :: Board -> Maybe Player

Winner : This function is to check whether any player has respective buttons in such a fashion that he can win the game. |This function is evoked after every single turn

countCells :: Cell -> Board -> Int

countCells : This function checks how many cells are empty in the board.

checkGameOver :: Game -> Game

checkGameOver : This checks if the game is over by checking the winning criteria or if its a tie |As this is invoked after every turn , if the winning criteria are not yet met then the game returns to its previous state.

playerTurn :: Game -> (Int, Int) -> Game

playerTurn : Depicts whose turn it is next to play the game |Also invokes checkGameOver function after every turn to see if the game has a winner |Otherwise it switches the player and goes on for the next turn

mousePositionAsCellCoordinate :: (Float, Float) -> (Int, Int)

mousePositionAsCellCoordinate : This function returns the row and column number of the cell when the player presses anywhere in the cell

slidedown :: Game -> (Int, Int) -> (Int, Int)

slidedown: This function is given the values of row and column number from mousePositionAsCellCoordinate |It takes the column number and checks if there is an empty cell beneath the row asked by player. |If there is a cell empty beneath the row specified by the player , the player's button will be allocated the first empty cell in the column starting from bottom.

transformGame :: Event -> Game -> Game

transformGame : This function is responsible to change the state of the game after every turn of player. |This function also checks if the game is complete or can still run after every turn.