Main

Description

Problem Statement:

Mr. X has a pair of integers (a,b), each drawn from the interval [2,100]. He gives the product, a * b, to a Mr. P and the sum, a + b, to Mr. S.

Mr. P and Mr. S have the following dialog:

Mr. P: "I don't know the numbers."

Mr. S: "I knew you didn't know. I don't know either."

Mr. P: "Now I know the numbers."

Mr. S: "Now I know them too."

Question: What are the numbers a and b?

Problem Solution:

The code below solves this problem. The answer is: (a,b) == (4,13).

Documentation

addPair :: (Integer, Integer) -> Integer

Add elements of a pair.

mulPair :: (Integer, Integer) -> Integer

Multiply elements of a pair.

formSums :: [Integer] -> [Integer]

Compute all possible sums of a list of integers, ls, without regard to the sum order. That is, in the resulting list of sums, only count one of (x + y) or (y + x).

formProds :: [Integer] -> [Integer]

Compute all possible products of a list of integers, ls, without regard to the product order. That is, in the resulting list of products, only count one of (x * y) or (y * x).

findPairSums :: (Ord b, Num b) => [b] -> b -> [(b, b)]

Find all the ways (discounting order) that s can be written as a sum from elements from ls.

findPairProds :: (Ord b, Num b) => [b] -> b -> [(b, b)]

Find all the ways (discounting order) that p can be written as a product from elements of ls.

unique :: Ord b => [b] -> [b]

Return the unique elements of the list ls in sorted order.

isSimpleNum :: (Ord b, Num b) => [b] -> b -> Bool

A simple number is one which has a unique factorization with respect to the list ls. That is, there is only one pair (discounting order) (e1,e2) in ls, such that p = e1 * e2.

getPossibleSumProds :: [Integer] -> [Integer] -> [(Integer, [Integer])]

Get "possible" sums/product combinations.

This means: For each element, sum, in sums, return only those that have

the property that each pair (x,y) such that x + y == sum

has the property that all x * y can be written as a product from ls

in more than one way.

This function returns a list of the form: [s, [pr]] where s is a sum made from ls + ls and [pr] are

all off the possible products that can be made of numbers from ls * ls which sum to s.

(We need this as Mr S. says that he doesn't know what the pair is.

The only way for Mr S. to know the pair is that there be only one one

pair of numbers that add to his sum. So we only look at sums that have more

than one combination and more than one product.)

getSums :: [Integer] -> Integer -> [Integer]

Get all possible sums that correspond to a product p.

getProds :: [Integer] -> Integer -> [Integer]

Get all possible products that correspond to a sum s.

sums :: [Integer] -> [Integer]

Define and solve problem with previous functions.

Get the unique sums formed by adding all pairs from the two lists.

prods :: [Integer] -> [Integer]

Get the unique products formed by multiplying all pairs from the two lists.

firstValidSums :: [Integer] -> [Integer] -> [Integer]

Get the first cut at valid info. A list of the form:

(s, ps), where s is a valid sum and ps are the possible products.

Mr S: "I knew you didn't"

Mr S has a sum which has the property that all possible ways of producing

it have the property that each of their products can be produced more than one way.

Get the first cut at the valid sums. These are the potential sums.

firstValidProds :: [Integer] -> [Integer] -> [Integer]

Get the first cut at the valid products. These are the potential products.

pSumValid :: Integer -> [Integer] -> [Integer] -> Bool

A filter, testing that for a given product there is one and only one sum from the list of potential sums (first potential list).

sPrdValid :: Integer -> [Integer] -> [Integer] -> Bool

A filter, testing that for a given sum there is one and only one product from the list of potential products (second potential list).

secondValidProds :: [Integer] -> p -> [Integer] -> [Integer] -> [Integer]

Second pass at valid products.

Mr. P: "Now I now the numbers."

Mr. P has a product for which there is only one way to produce the corresponding sum

(from the latest list of potential sums).

solutions :: [Integer] -> [(Integer, Integer)]

Find all the pairs (a,b) such that a + b is one of the sums, sums,

and a * b is one of the products, prds.

Question: "What are the numbers a and b?"

From what we know of potential sums and potential products, we find all possible pairs

which have sums in the list of potential sums and products in the list of potential products.

main :: IO ()

Main entry point. Return the solution pairs.