Writer

Documentation

data Writer log a

The data type that represents a writer

Constructors

Writer log a

tell :: Monoid log => log -> Writer log ()

Adds a value to the log

return :: Monoid log => a -> Writer log a

andThen :: Monoid log => Writer log a -> (a -> Writer log b) -> Writer log b

Executes the first writer and then the second writer with the a from the first, with side effects

andThen' :: Monoid log => Writer log a -> Writer log b -> Writer log b

Executes the first writer, throws away its result (but not its logs), then evaluates the second parser