Main

Description

O intuito do projeto é fazer um jogo de Tetris em Haskell utilizando a biblioteca Gloss.

Documentation

listaPecas :: IO [Tetramino]

Lista de Tetraminós gerada aleatóriamente.

mundoInicial :: [Tetramino] -> Mundo

Mundo inicial do jogo que contém apenas a primeira peça.

joga :: Event -> Mundo -> Mundo

Se possível, movimenta a peça de acordo com a tecla pressionada, atualizando os blocos do mundo e da peça ativa.

podeMover :: Mundo -> Direcao -> Bool

Verifica se um movimento pode ser feito dado um mundo e uma direção.

move :: Direcao -> Peca -> Peca

Move uma peça para uma direção dada, atualizando os blocos da mesma.

voltaLimite :: Coord -> Float

Devolve por quanto que um bloco ultrapassou os limites da tela.

moveFundo :: Mundo -> Peca

Movimenta a peça direto para o fundo da tela, movendo pra baixo recursivamente.

atualizaMundo :: Float -> Mundo -> Mundo

Atualiza o mundo.

limpaFileira :: Mundo -> Mundo

Limpa as fileiras completas do mundo recursivamente (de acordo com os blocos da peça), atualizando o número de fileiras completas.

atualizaPont :: Int -> Int

Sistema de pontuação do jogo baseada na quantidade de fileiras completas de uma vez.

main :: IO ()